Initial import of markdown2social sources
The markdown2social utility converts simple Markdown documents to Google+
posts ready to be shared on the social network.  There is room for writing
output plugins for other target sites, but at the moment only Google+ is
supported.

This is the open-source release of a tool I originally wrote internally at
Google called md2gplus, with edits to adapt it to the public Google Python
coding style.  I have used this tool a bunch of times to draft internal
posts as Markdown documents that are later published as Google+ posts for
greater engagement.

This is not an official Google product.